+++ lib/src/render/blocks/code.dart
@@ -46,8 +46,8 @@ class BlockPainter$Code with SelectableTextBlock implements BlockPainter {
     MarkdownThemeData theme,
   ) {
     final baseStyle = theme.textStyle.copyWith(
-      fontFamily: kMonospaceFontFamily,
-      fontFamilyFallback: kMonospaceFontFamilyFallback,
+      fontFamily: theme.effectiveMonospaceFontFamily,
+      fontFamilyFallback: theme.effectiveMonospaceFontFamilyFallback,
       fontSize: theme.textStyle.fontSize ?? kDefaultFontSize,
     );
     final highlighter = theme.highlighter;